cleaning up code and linking, sort of
- Removed a pointless rotation function.
- Removed a pointless check.
- Added some missing documentation (theres still more missing than i
would like, I'll have to take a day sometime soon and just double check
all my documentation).
- Linking now works, sometimes. the method will need cleaning up and
better compensation for different circumstances, but it's a really good
starting point.

+++ src/main/java/trains/models/RenderEntity.java
@@ -207,10 +207,10 @@ public class RenderEntity extends Render {
             float randomZ;
             float colorOffset;
             Random rand = new Random();
-            float[] direction;
+            double[] direction;
             for (float[] smoke : entity.getSmokeOffset()) {
                 for (int i = 0; i < smoke[4]; i++) {
-                    direction = RailUtility.rotatePoint(new float[]{smoke[0], smoke[1], smoke[2]}, entity.rotationPitch, entity.rotationYaw, 0);
+                    direction = RailUtility.rotatePoint(new double[]{smoke[0], smoke[1], smoke[2]}, entity.rotationPitch, entity.rotationYaw, 0);
                     direction[0] += entity.posX;
                     direction[1] += entity.posY;
                     direction[2] += entity.posZ;
